refactor:convert type of available position rate calculation

+++ src/main/java/com/parkinglot/ParkingLot.java
@@ -29,6 +29,6 @@ public class ParkingLot {
     }
 
     public double getAvailablePositionRate (){
-        return getNumberOfEmptyPositions() / MAX_CAPACITY;
+        return (double) getNumberOfEmptyPositions() / MAX_CAPACITY;
     }
 }
